conv/conv_test.go
package conv

import (
	"reflect"
	"testing"
)

func TestFahrenheitToCelsius(t *testing.T) {
	type test struct {
		  input float64
		  want  float64
	}

	tests := []test{
		  {input: 134, want: 56.67},
	}

	for _, tc := range tests {
		  got := FahrenheitToCelsius(tc.input)
		  if !reflect.DeepEqual(tc.want, got) {
				 t.Errorf("expected: %v, got: %v",
										 tc.want, got)
		  }
	}
}

func TestCelsiusToFahrenheit(t *testing.T) {
	type test struct {
		  input float64
		  want  float64
	}

	tests := []test{
		  {input: 47, want: 116.6},
	}

	for _, tc := range tests {
		  got := CelsiusToFahrenheit(tc.input)
		  if !reflect.DeepEqual(tc.want, got) {
				 t.Errorf("expected: %v, got: %v",
										 tc.want, got)
		  }
	}
}

func TestCelsiusToKelvin(t *testing.T) {
	type test struct {
		  input float64
		  want  float64
	}

	tests := []test{
		  {input: -40, want: 233.15},
	}

	for _, tc := range tests {
		  got := CelsiusToKelvin(tc.input)
		  if !reflect.DeepEqual(tc.want, got) {
				 t.Errorf("expected: %v, got: %v",
										 tc.want, got)
		  }
	}
}

func TestKelvinToCelsius(t *testing.T) {
	type test struct {
		  input float64
		  want  float64
	}

	tests := []test{
		  {input: 400, want: 126.85},
	}

	for _, tc := range tests {
		  got := KelvinToCelsius(tc.input)
		  if !reflect.DeepEqual(tc.want, got) {
				 t.Errorf("expected: %v, got: %v",
										 tc.want, got)
		  }
	}
}

func TestKelvinToFahrenheit(t *testing.T) {
	type test struct {
		  input float64
		  want  float64
	}

	tests := []test{
		  {input: 300, want: 80.33},
	}

	for _, tc := range tests {
		  got := KelvinToFahrenheit(tc.input)
		  if !reflect.DeepEqual(tc.want, got) {
				 t.Errorf("expected: %v, got: %v",
										 tc.want, got)
		  }
	}
}

func TestFahrenheitToKelvin(t *testing.T) {
	type test struct {
		  input float64
		  want  float64
	}

	tests := []test{
		  {input: 9, want: 260.37},
	}

	for _, tc := range tests {
		  got := FahrenheitToKelvin(tc.input)
		  if !reflect.DeepEqual(tc.want, got) {
				 t.Errorf("expected: %v, got: %v",
										 tc.want, got)
		  }
	}
}